extracting fetch call into utils
