+++ frontend/src/components/Home/Home.js
@@ -1,6 +1,7 @@
 import React, { useEffect } from "react";
 
 import { useAuth0 } from "../../react-auth0-spa";
+import addAccessToken from "../../utils/commonRequests.js";
 
 import "./Home.css";
 
@@ -13,23 +14,10 @@ let gridItems;
 
 function Home() {
   const { isAuthenticated, user } = useAuth0();
-
   // should be fetching on load of the Home page and NOT the Navigation bar (Website component)
   // that was my mistake earlier
   useEffect(() => {
-    if (isAuthenticated && user) {
-      const body = {
-        uid: user.sub
-      };
-      fetch("http://localhost:3001/setAccessToken", {
-        method: "POST",
-        headers: {
-          Accept: "text/html",
-          "Content-Type": "application/json"
-        },
-        body: JSON.stringify(body)
-      });
-    }
+    addAccessToken(isAuthenticated, user);
   });
 
   gridItems = [];
